Double_ended_queue.c
C program to implement the Double ended Queue Data structure. This program does the job of both Input as well as output restricted Dequeue.
If you want an Input restricted Dequeue, then, in the insert function, pass 0 and in the delete function pass 1 as the 3rd parameter.
Vice-versa for Output Restricted Dequeue.