Create Digital_root.cpp
There is a problem, in which we have to keep summing up the digits, till the sum becomes a single-digit. There is an iterative process, but it will lead to time-exceed error. If we go for the recursive approach, then there will be memory leakage/Stack Overflow.

But there is a concept to solve it. The problem is based on Digital Root. In this problem, if the number is divisible by 9, its digital root will be 9 else the remainder when that number is divided by 9.