Added code for Happy Number in Java..
The happy number can be defined as a number which will yield 1 when it is replaced by the sum of the square of its digits repeatedly. If resulting sum is equal to 1 then, given number is a happy number. If the sum is equal to 4 then, the number is an unhappy number.